feat: add GUI

+++ src/cssom.rs
@@ -102,11 +102,11 @@ pub struct Color {
   pub r: u8,
   pub g: u8,
   pub b: u8,
-  pub a: u8,
+  pub a: f32,
 }
 
 impl Color {
-  pub fn new(r: u8, g: u8, b: u8, a: u8) -> Color {
+  pub fn new(r: u8, g: u8, b: u8, a: f32) -> Color {
     Color { r, g, b, a }
   }
 }
